requirements.txt
Flask==2.3.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Authlib==1.2.1
Werkzeug==2.3.0
requests==2.31.0
Pillow==10.0.0
pytesseract==0.3.10
joblib==1.3.2
scikit-learn==1.3.0
pandas==2.0.3
spacy==3.6.0
psycopg2-binary==2.9.7
